// core/database/interfaces/contact.ts
export interface Contact {
  companyName: string;
  brandName?: string;
  ruc: string;
  establishmentType: string;
  foodType: string;
  address?: string;
  phone?: string;
  email?: string;
  contactName: string;
  contactLastName: string;
  contactId: string;
  contactPosition: string;
  contactPhone: string;
  contactEmail: string;
}